[Fix] modify 무한루프 해결하지 못하고 상수 선언으로 당장은 해결
