+++ client/src/configs/Listpage/config.tsx
@@ -1,0 +1,9 @@
+import { SearchType } from 'Types/Types';
+
+export const search: SearchType = {
+  teacherName: '',
+  subject: [],
+  regions: [],
+  size: 0,
+  page: 0,
+};
