+++ client/src/redux/slice/ListPageSlice.tsx
@@ -1,5 +1,5 @@
 import { PayloadAction, createSlice } from '@reduxjs/toolkit';
-import { ListPageType, SearchType } from 'Types/Types';
+import { ListPageType } from 'Types/Types';
 import { RootState } from 'redux/store';
 import { getData } from 'redux/thunk/ListPageThunk';
 
@@ -13,7 +13,6 @@ type initialStateType = {
       totalElements: number;
       totalPages: number;
     };
-    search: SearchType;
   };
 };
 
@@ -26,13 +25,6 @@ const initialState: initialStateType = {
       size: 0,
       totalElements: 0,
       totalPages: 0,
-    },
-    search: {
-      teacherName: '',
-      subject: [],
-      regions: [],
-      size: 0,
-      page: 0,
     },
   },
 };
